# swagger-ui/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "SEMOSS",
    "version" : "0.0.1-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "https://workshop.cfg.deloitte.com/cfg-ai-demo/Monolith",
    "description" : "Demo Endpoint URL"
  }, {
    "url" : "http://localhost:8080/Monolith",
    "description" : "Local Development"
  } ],
...
  "paths" : {
    "/app-{databaseId}/appImage/download" : {
      "get" : {
        "operationId" : "imageDownload_1",
        "parameters" : [ {
          "name" : "databaseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/octet-stream" : { },
              "application/svg+xml" : { }
            }
          }
        },
        "deprecated" : true,
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
    "/app-{databaseId}/updateSmssFile" : {
      "post" : {
        "operationId" : "updateSmssFile_1",
        "parameters" : [ {
          "name" : "databaseId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json;charset=utf-8" : { }
            }
          }
        },
        "deprecated" : true,
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
...
    "/e-{engineId}/image/download" : {
      "get" : {
        "operationId" : "imageDownload",
        "parameters" : [ {
          "name" : "engineId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/octet-stream" : { },
              "application/svg+xml" : { }
            }
          }
        }
      }
    },
    "/e-{engineId}/updateSmssFile" : {
      "post" : {
        "operationId" : "updateSmssFile",
        "parameters" : [ {
          "name" : "engineId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json;charset=utf-8" : { }
            }
          }
        }
      }
    },
...
    "/engine/runPixel" : {
      "post" : {
        "operationId" : "runPixelSync",
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json;charset=utf-8" : { }
            }
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
...
    "/project-{projectId}/updateSmssFile" : {
      "post" : {
        "operationId" : "updateSmssFile_2",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json;charset=utf-8" : { }
            }
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
...
  }
}